backend/requirements.txt
albucore==0.0.24
albumentations==2.0.8
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
billiard==4.2.2
cachetools==5.5.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.6
easydict==1.13
fastapi==0.109.1
flatbuffers==25.9.23
fonttools==4.60.1
google-api-core==2.27.0
google-api-python-client==2.90.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.71.0
greenlet==3.2.4
h11==0.16.0
httplib2==0.31.0
humanfriendly==10.0
idna==3.11
imageio==2.37.0
insightface==0.7.3
joblib==1.5.2
kiwisolver==1.4.9
kombu==5.5.4
lazy_loader==0.4
matplotlib==3.10.7
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
oauthlib==3.3.1
onnx==1.19.1
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
packaging==25.0
pgvector==0.4.1
pillow==12.0.0
prettytable==3.16.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.3
redis==7.0.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.35.1
stringzilla==4.2.1
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
alembic
pgvector
